Set train error correction hidden size to 768
We need this for another baseline. This baseline experiment is preparation
for the experiment where we'll use BERT vectors to initialize the hidden
state of the encoder of the seq2seq model.

We expect that a larger hidden sizes leads to longer training, so to be
on the safe side, the number of epochs was also increased.